# src/seisspark_service/openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/pipelines":{"get":{"tags":["pipelines"],"summary":"Get Pipelines","operationId":"get_pipelines","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Pipelines Api V1 Pipelines Get","type":"array","items":{"$ref":"#/components/schemas/PipelineInfo"}}}}}}},"post":{"tags":["pipelines"],"summary":"Create Pipelines","operationId":"create_pipeline","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePipelineRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PipelineInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pipelines/{pipeline_id}":{"get":{"tags":["pipelines"],"summary":"Get Pipeline","operationId":"get_pipeline","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PipelineInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["pipelines"],"summary":"Delete Pipeline","operationId":"delete_pipeline","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pipelines/{pipeline_id}/modules":{"get":{"tags":["pipelines"],"summary":"Get Pipeline Modules","operationId":"get_pipeline_modules","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Pipeline Modules Api V1 Pipelines  Pipeline Id  Modules Get","type":"array","items":{"$ref":"#/components/schemas/ModuleInfo"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["pipelines"],"summary":"Move Pipeline Module","operationId":"move_pipeline_module","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoveModuleRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["pipelines"],"summary":"Create Pipeline Module","operationId":"create_pipeline_module","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateModuleRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModuleDescription"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pipelines/{pipeline_id}/modules/{module_id}":{"delete":{"tags":["pipelines"],"summary":"Delete Pipeline Module","operationId":"delete_pipeline_module","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"},{"required":true,"schema":{"title":"Module Id","type":"string"},"name":"module_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pipelines/{pipeline_id}/modules/{module_id}/parameters":{"get":{"tags":["pipelines"],"summary":"Get Pipeline Module Parameters","operationId":"get_pipeline_module_parameters","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"},{"required":true,"schema":{"title":"Module Id","type":"string"},"name":"module_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["pipelines"],"summary":"Set Pipeline Module Parameters","operationId":"set_pipeline_module_parameters","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"},{"required":true,"schema":{"title":"Module Id","type":"string"},"name":"module_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Parameters","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pipelines/{pipeline_id}/modules/{module_id}/schema":{"get":{"tags":["pipelines"],"summary":"Get Pipeline Module Schema","operationId":"get_pipeline_module_schema","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"},{"required":true,"schema":{"title":"Module Id","type":"string"},"name":"module_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pipelines/{pipeline_id}/modules/{module_id}/keys":{"get":{"tags":["pipelines"],"summary":"Get Pipeline Module Data Info","operationId":"get_pipeline_module_data_info","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"},{"required":true,"schema":{"title":"Module Id","type":"string"},"name":"module_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/pipelines/{pipeline_id}/modules/{module_id}/data/{key}":{"get":{"tags":["pipelines"],"summary":"Get Pipeline Module Data","operationId":"get_pipeline_module_data","parameters":[{"required":true,"schema":{"title":"Pipeline Id","type":"string"},"name":"pipeline_id","in":"path"},{"required":true,"schema":{"title":"Module Id","type":"string"},"name":"module_id","in":"path"},{"required":true,"schema":{"title":"Key","type":"integer"},"name":"key","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/modules":{"get":{"tags":["modules"],"summary":"Get Modules","operationId":"get_modules","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Modules Api V1 Modules Get","type":"array","items":{"type":"string"}}}}}}}},"/api/v1/modules/{module_type}":{"get":{"tags":["modules"],"summary":"Get Module Schema","operationId":"get_module_schema","parameters":[{"required":true,"schema":{"title":"Module Type","type":"string"},"name":"module_type","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CreateModuleRequest":{"title":"CreateModuleRequest","required":["module_type"],"type":"object","properties":{"module_type":{"title":"Module Type","type":"string"},"name":{"title":"Name","type":"string"},"prev_module_id":{"title":"Prev Module Id","type":"string"}}},"CreatePipelineRequest":{"title":"CreatePipelineRequest","required":["name"],"type":"object","properties":{"name":{"title":"Name","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ModuleDescription":{"title":"ModuleDescription","required":["id","name","params_schema"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"params_schema":{"title":"Params Schema","type":"object"}}},"ModuleInfo":{"title":"ModuleInfo","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"MoveModuleRequest":{"title":"MoveModuleRequest","required":["module_id"],"type":"object","properties":{"module_id":{"title":"Module Id","type":"string"},"prev_module_id":{"title":"Prev Module Id","type":"string"}}},"PipelineInfo":{"title":"PipelineInfo","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}